requirements.txt
chromadb
numpy
open_clip_torch
pillow
sentencepiece
sentencepiece
transformers
torch